DB + Refactor of Telnet Opt Parse + ReceiveLine
Added a sqlite database with a key-value table and two functions to use
from within lua. This key value pair is cached by the server itself.

Added a toRecv buffer for clients and moved parsing of received bytes to
a similar loop that parses toSend. This basically means that telnet opts
and lines can split packets.

As part of the above added the DMUD_ReceivedLine callback, this only
collects the 127 ASCII characters excluding telnet opts and new lines.